refactor: useSpecialtyFilter 훅 추가 및 Home 컴포넌트에서 전문 분야 상태 관리 개선 #1047
